+++ HISTORY.md
@@ -9,7 +9,8 @@
 * symbols: added a single "gear" symbol, changed "gears" to a pair of interlocking gears
 * gui: changed (back) symbol for the settings button to a single gear symbol
 * symbols: refactored "gears" symbol rendering 
-* symbols: refactored singe "gear" symbol rendering 
+* symbols: refactored singe "gear" symbol rendering
+* symbols: added new "gearsAnimated" symbol
 
 ### 2026-07-24
 * new dev version